LPS-195477 Merger class DataStorageRegistry with the class DataRecordResourceImpl
